+++ Basestation/Pad.h
@@ -8,14 +8,16 @@
 class ChargePad
 {
   public:
-    ChargePad( uint8_t pinColorSensorSDA, 
-                      uint8_t pinColorSensorSCL, 
-                      uint8_t pinColorSensorLED,
-                      uint8_t pinChargeEnable,
-                      uint8_t pinCurrentSense,
-                      uint8_t pinVoltageSense,
-                      uint16_t voltageSenseResistanceR1,
-                      uint16_t voltageSenseResistanceR2);
+    ChargePad( 
+                uint8_t pinColorSensorSDA, 
+                uint8_t pinColorSensorSCL, 
+                uint8_t pinColorSensorLED,
+                uint8_t pinChargeEnable,
+                uint8_t pinCurrentSense,
+                uint8_t pinVoltageSense,
+                uint32_t voltageSenseResistanceR1,
+                uint32_t voltageSenseResistanceR2
+              );
     
     
     void calibrateColorSensor( void );
